Partially resolves issue #821
This commit fixes issues with sorting by price, but does not address category sorting preferences being ignored (I'm not sure where those settings can be found).

Changing search order via the select menu functions correctly (tested name and price, both ASC and DESC) except that the default sort for the initial search appears to be by relevance while the dropdown menu shows 'Name (A-Z)' as the currently selected method. This means that the initial search results appear to be mis-ordered, but it is merely a mis-display of the selected option.

Keying off of `G.product_id` is more correct as it only uses the group pricing if the group itself exists, whereas `G.sale_price` can be null both when the group doesn't exist and when the group doesn't have a sale price.

Furthermore, `sale_price` can have a value of either `null` or `0.00`, both of which should be considered non-values for purposes of sorting by price.